% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{CleanText}
\alias{CleanText}
\title{CleanText}
\usage{
CleanText(text, pdfExtractor = F, newLine = T, dashLine = F,
  lowerCase = F)
}
\arguments{
\item{text}{text}

\item{pdfExtractor}{pdfExtractor}

\item{newLine}{newLine}

\item{dashLine}{dashLine}

\item{lowerCase}{lowerCase}
}
\value{
CleanText
}
\description{
CleanText
}
