% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{ConvertPdftoText}
\alias{ConvertPdftoText}
\title{ConvertPdftoText}
\usage{
ConvertPdftoText(pdfLink, ignoreExistingTextFile = FALSE,
  convertSoftware = "\\"pdftotext\\"")
}
\arguments{
\item{pdfLink}{pdfLink}

\item{ignoreExistingTextFile}{ignore ExistingTextFile}

\item{convertSoftware}{convertSoftware}
}
\value{
status
}
\description{
Convert pdf to text
}
