% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{CountPattern}
\alias{CountPattern}
\title{CountPattern}
\usage{
CountPattern(text, pattern, ignoreCase = T)
}
\arguments{
\item{text}{Character.}

\item{pattern}{Character.}

\item{ignoreCase}{Boolean.}
}
\value{
number of pattern found in text
}
\description{
The function count any pattern, regular expression acceptable, in text. The function uses perl
}
