% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{CountPatternInPar}
\alias{CountPatternInPar}
\title{CountPatternInPar}
\usage{
CountPatternInPar(myStudies = myStudies, myDictionary = myDictionary,
  textSearchingHeaders = textSearchingHeaders,
  linkSearchHeaders = linkSearchHeaders,
  dictionaryNameHeader = dictionaryNameHeader,
  dictionaryRegexHeader = dictionaryRegexHeader, ignoreCase = ignoreCase)
}
\arguments{
\item{myStudies}{myStudies}

\item{myDictionary}{myDictionary}

\item{textSearchingHeaders}{textSearchingHeaders}

\item{linkSearchHeaders}{linkSearchHeaders}

\item{dictionaryNameHeader}{dictionaryNameHeader}

\item{dictionaryRegexHeader}{dictionaryRegexHeader}

\item{ignoreCase}{ignoreCase}
}
\value{
frequency
}
\description{
Count Pattern In Par
}
