% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{CountPatternOverMatrix}
\alias{CountPatternOverMatrix}
\title{CountPatternOverMatrix}
\usage{
CountPatternOverMatrix(text, pattern, margin = margin,
  ignoreCase = ignoreCase)
}
\arguments{
\item{pattern}{Character string containing a regular expression to be matched in the given character vector.
Coerced by as.character to a character string if possible.
If a character vector of length 2 or more is supplied, the first element is used with a warning. Missing values are allowed.
If the data frame contains more than one column, then the only first column is used.}

\item{ignoreCase}{Boolean}

\item{fullText}{A list, arry or data frame of character vectors where matches are sought, or a list or array of objects which can be coerced by as.character to a character vector. Long vectors are supported.}
}
\value{
a list of integers showing the number frequency of the pattern match over columns and over rows
}
\description{
Count Pattern Over Matrix
}
