% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{CountTermsInStudies}
\alias{CountTermsInStudies}
\title{CountTermsInStudies}
\usage{
CountTermsInStudies(searchingData = NULL, dictionary = NULL,
  searchingHeaders = c("Title", "Abstract"),
  textSearchingHeaders = c("Title", "Abstract"),
  linkSearchHeaders = c("PdfRelativePath"), dictionaryNameHeader = "Name",
  dictionaryRegexHeader = "Regex", ignoreCase = "T")
}
\arguments{
\item{searchingData}{Either a dataset or a link to the dataset to search from}

\item{dictionary}{Either a dictionary dataset, or a link to the dictionary dataset to run the function on.
It should consist two columns: name of the term and search string of the term. Regular expression (Perl) is accepted for the search string.
If there is only one column, that column will be used both as name and regular expression.}

\item{textSearchingHeaders}{A list of the headers of the columns to search from. A list of character. Default value is c('Title', 'Abstract', 'FullText')}

\item{dictionaryNameHeader}{The header string of name column in dictionary}

\item{dictionaryRegexHeader}{The header string of regular expression column in dictionary}

\item{pdfLinkSearchHeaders}{A list of the headers of the columns to convert, read and search from. A list of character. Default value is c('Title', 'Abstract', 'FullText')}

\item{textLinkSearchHeaders}{A list of the headers of the columns to read and search from. A list of character. Default value is c('Title', 'Abstract', 'FullText')}
}
\value{
A data frame with result of the dictionary search. One column for each term in the dictionary, with the name of the term as header.
}
\description{
CountTermsInStudies
}
