% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{ExtractColumns}
\alias{ExtractColumns}
\title{ExtractColumns}
\usage{
ExtractColumns(df, columnNames)
}
\arguments{
\item{df}{a data frame}

\item{columnNames}{a list of character object}
}
\value{
a data frame with column names matching columnNames. If not mathing is found, NULL is returned.
}
\description{
Extract the columns in the df matching the
}
