% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{ReadFullText}
\alias{ReadFullText}
\title{ReadFullText}
\usage{
ReadFullText(txtFileName)
}
\arguments{
\item{txtFileName}{text file name}
}
\value{
full text
}
\description{
Read full text
}
