% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{RiskOfBiasIdentification}
\alias{RiskOfBiasIdentification}
\title{RiskOfBiasIdentification}
\usage{
RiskOfBiasIdentification(searchingData)
}
\arguments{
\item{searchingData}{}
}
\value{
result flags
}
\description{
# Identify Risk of Bias in myData$cleanText
}
