% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{CountPatternInPar}
\alias{CountPatternInPar}
\title{CountPatternInPar}
\usage{
CountPatternInPar(myStudies = NULL, myDictionary,
  textSearchingHeaders = "", linkSearchHeaders = "",
  dictionaryNameHeader = "Name", dictionaryRegexHeader = "Regex",
  ignoreCase = TRUE, ignoreExistingTextFile = TRUE,
  conversionSoftware = "pdftotext")
}
\arguments{
\item{myStudies}{myStudies}

\item{myDictionary}{myDictionary}

\item{textSearchingHeaders}{textSearchingHeaders}

\item{linkSearchHeaders}{linkSearchHeaders}

\item{dictionaryNameHeader}{dictionaryNameHeader}

\item{dictionaryRegexHeader}{dictionaryRegexHeader}

\item{ignoreCase}{ignoreCase}

\item{ignoreExistingTextFile}{ignoreExistingTextFile}

\item{conversionSoftware}{Software used to covert pdf to text. Default value is 'pdftotext'}
}
\value{
frequency
}
\description{
Count Pattern In Par
}
