% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{CountPatternInPar}
\alias{CountPatternInPar}
\title{CountPatternInPar}
\usage{
CountPatternInPar(
  myStudies = NULL,
  myDictionary,
  textSearchingHeaders = "",
  linkSearchHeaders = "",
  dictionaryNameHeader = "Name",
  dictionaryRegexHeader = "Regex",
  ignoreCase = TRUE,
  ignoreExistingTextFile = TRUE,
  cutIntro = FALSE,
  cutRefs = FALSE,
  extractStrings = FALSE,
  conversionSoftware = "pdftotext"
)
}
\arguments{
\item{myStudies}{myStudies}

\item{myDictionary}{myDictionary}

\item{textSearchingHeaders}{textSearchingHeaders}

\item{linkSearchHeaders}{linkSearchHeaders}

\item{dictionaryNameHeader}{dictionaryNameHeader}

\item{dictionaryRegexHeader}{dictionaryRegexHeader}

\item{ignoreCase}{boolean to decide whether to ignore the case in searching the content in dictionary in the searchingData or not}

\item{ignoreExistingTextFile}{ignoreExistingTextFile}

\item{extractStrings}{= boolean variable to indicate whether matching strings from full text should be extracted}

\item{conversionSoftware}{Software used to covert pdf to text. Default value is 'pdftotext'}
}
\value{
frequency
}
\description{
Count Pattern In Par
}
