% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{ReadLink}
\alias{ReadLink}
\title{ReadLink}
\usage{
ReadLink(link)
}
\arguments{
\item{link}{read full text from link}

\item{pdfLink}{pdfLink}

\item{ignoreExistingTextFile}{ignore ExistingTextFile}

\item{convertSoftware}{convertSoftware}
}
\value{
linkstatus and fulltext

status
}
\description{
Read Link
}
