% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAnnotationFunctions.R
\name{RiskOfBiasIdentification}
\alias{RiskOfBiasIdentification}
\title{RiskOfBiasIdentification}
\usage{
RiskOfBiasIdentification(searchingData)
}
\arguments{
\item{searchingData}{data to search from. Either a link to the file or a data frame or a matrix}
}
\value{
result flags
}
\description{
Identify Risk of Bias in myData$cleanText
}
